requirements.txt
neo4j
numpy
trimesh
networkx
scipy
python-dotenv